Breaking message audit and saga audit into separate features to allow use in endpoints that don't have any sagas
